+++ src/index.ts
@@ -1,4 +1,8 @@
-export type { AccessDialog, HandleAccessRequest } from "./access-dialog.js";
+export type {
+  AccessDialog,
+  AccessDialogResult,
+  HandleAccessRequest,
+} from "./access-dialog.js";
 export { createDelegatedPermissions } from "./delegated-permissions.js";
 export type {
   IsDelegateSelected,
@@ -9,6 +13,7 @@ export { createPermissionObserver } from "./permission-observer.js";
 export type { PermissionObserver } from "./permission-observer.js";
 export { createPermissionStore } from "./permission-store.js";
 export type {
+  IsMatchingDescriptor,
   PermissionAccessState,
   PermissionAccessStatus,
   PermissionStore,
@@ -16,4 +21,4 @@ export type {
 } from "./permission-store.js";
 export { createPermissions } from "./permissions.js";
 export { createUser } from "./user.js";
-export type { User } from "./user.js";
+export type { AccessRequest, User } from "./user.js";